Change the import path for AzimuthalIntegrator
