+++ src/pyFAI/integrator/meson.build
@@ -1,6 +1,7 @@
 py.install_sources(
    ['__init__.py',
     'azimuthal.py',
+    'common.py',
     'fiber.py',
     'load_engines.py'],
   pure: false,   # Will be installed next to binaries
