+++ src/pyFAI/load_integrators.py
@@ -30,7 +30,7 @@ __author__ = "Jérôme Kieffer"
 __contact__ = "Jerome.Kieffer@ESRF.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "09/10/2024"
+__date__ = "10/10/2024"
 __status__ = "stable"
 __docformat__ = 'restructuredtext'
 
@@ -40,4 +40,8 @@ Compatibility layer, the code has been moved to pyFAI.integrator.load
 09/10/2024
 """
 
+# TODO put in place a deprecation warning when importing this module
+
 from .integrator.load_engines import *
+from .utils.decorators import deprecated_warning
+deprecated_warning("Module", "pyFAI.load_integrators", replacement="pyFAI.integrator.load_engines", since_version="2024.10", only_once=False)
