+++ src/pyFAI/opencl/test/test_ocl_azim_csr.py
@@ -44,7 +44,7 @@ if ocl:
     import pyopencl.array
 from ...test.utilstest import UtilsTest
 from silx.opencl.common import _measure_workgroup_size
-from ...azimuthalIntegrator import AzimuthalIntegrator
+from ...integrator.azimuthal import AzimuthalIntegrator
 from ...method_registry import IntegrationMethod
 from scipy.ndimage import gaussian_filter1d
 logger = logging.getLogger(__name__)
